+++ client/src/components/Inventory/Inventory.js
