+++ client/src/components/Inventory/Stocks.js
